CIT plugin change_one_value fix
When the chosen value can't be changed, bacause of some constrain. The function randomly choose diferent valueWhen the chosen value couldn't be changed, because of some constraints. The function randomly chose different value, which is not expected behavior. Now when this event occurs, the function raises ValueError exception and it's handled.

Signed-off-by: Jan Richter <jarichte@redhat.com>