Runner / Test: add support for SETUP phase specific timeout
This allows a test to define the timeout, or maximum amount of
time for completion, for the SETUP phase of a test.  This is currently
implemented only for INSTRUMENTED tests, that actually report this
phase.

If a test class defines a "setup_timeout" attribute, the runner will
differentiate between the "timeout" attribute/parameter.  See the
accompanying example test (long_setup.py) for an example.

Signed-off-by: Cleber Rosa <crosa@redhat.com>