Drop test that locally build docs
In my experience, a very small percentage of the changes actually
introduce Sphinx based documentation.  Say, a feature may be composed
of 10 patches, and at the end of the series, the feature is
documented.

Requiring the documentation to be build all the time takes a toll on
both local and remote systems CI systems.  Let's make the testing
slimmer, and rely on readthedocs.org to tell if any issues are
introduced.

Instead, let's deal with eventual issues before a release is cut.

Signed-off-by: Cleber Rosa <crosa@redhat.com>