Revert "Test: report state on each phase"
Some aspects of the test runner, including how it applies timeouts and
the specific type of timeout, relies on the absence of a status to
determine that a test process may or may be alive.

At least one test, test_kill_stopped_sleep fom
selftests.functional.test_basic.RunnerSimpleTest will fail because the
runner applies a different type of timeout on TestStatus.finish().

This is far from ideal, but it's how a lot of the current runner
implementation works, and given the current timing, it's better
to keep the behavior consistent, tests passing, and disrupt the
internals on a new development cycle.

This reverts commit ac776472205690423cb00f94723684c570ced287.

Signed-off-by: Cleber Rosa <crosa@redhat.com>