Move extension manager with disable/order support to its own module
This class is not really a dispatcher designed to be used standalone,
but a extension manager with additional features that can be used as
the basis for real dispatchers.

If it didn't depend on the settings module, it'd be a good candidate
for living in the avocado.core.extension_manager module, but to avoid
circular dependencies there, it needs to live elsewhere.  With those
things in mind, let's put it in its own module.

Signed-off-by: Cleber Rosa <crosa@redhat.com>