Job API enabler: switch from Namespace to dict for job config
A argparse.Namespace instance has been used so far for the
configuration of virtually all job (and plugins) aspects.  The problem
is that argparse.Namespace is really intended as the result of the
parsing of the command line, and it's a lot less flexible and less
natural for developers than a Python dictionary.

Given that we want to let users write their own jobs, it makes
sense to use Python native data types for that.  For instance,
this is a trivial job that can be written with this change:

   from avocado.core.job import Job
   config = {'reference': ['examples/tests/passtest.py:PassTest.test',
                           'examples/tests/passtest.py:PassTest.test']}
   with Job(config) as j:
       j.run()

Instea of:

   import argparse
   from avocado.core.job import Job
   config = argparse.Namespace()
   setattr(config, 'reference', ['examples/tests/passtest.py:PassTest.test',
                                  'examples/tests/passtest.py:PassTest.test'])
   with Job(config) as j:
       j.run()

As soon as the command line parser is done parsing the command line
arguments (see avocado.core.parser.Parser.finish()) args will be
become a dictionary.

In the future, we may want to rename "args" to something like config
or job_config, so that it's more obvious what it's about and doesn't
get confused with say positional args.

Signed-off-by: Cleber Rosa <crosa@redhat.com>