Job API enabler: switch from Namespace to dict for job config
A argparse.Namespace instance has been used so far for the
configuration of virtually all job (and plugins) aspects.  The problem
is that argparse.Namespace is really intended as the result of the
parsing of the command line, and it's a lot less flexible and less
natural for developers than a Python dictionary.

Given that we want to let users write their own jobs, it makes
sense to use Python native data types for that.  For instance,
this is a trivial job that can be written with this change:

   from avocado.core.job import Job
   config = {'reference': ['examples/tests/passtest.py:PassTest.test',
                           'examples/tests/passtest.py:PassTest.test']}
   with Job(config) as j:
       j.run()

Instead of:

   import argparse
   from avocado.core.job import Job
   config = argparse.Namespace()
   setattr(config, 'reference', ['examples/tests/passtest.py:PassTest.test',
                                  'examples/tests/passtest.py:PassTest.test'])
   with Job(config) as j:
       j.run()

As soon as the command line parser is done parsing the command line
arguments (see avocado.core.parser.Parser.finish()) args will be
become a dictionary.

The attribute has also been renamed from args to config, as it more
accurately describes the configuration that can be suplied to a job.
It makes it clearer that this dictionary doesn't have a dependency on
the command line parser.

It's expected that the job configuration can be loaded by other from
other sources (JSON, YAML) that can be converted to a dictionary, or
even expressed as a instance in the job definition.

Signed-off-by: Cleber Rosa <crosa@redhat.com>