Improve the HTML report
Bring back some features from autotest report, namely the
ability to display all records in the table and an extra dropdown
to filter records by status.

Also, to avoid breaking the table layout when we have long
values we truncate those and add an ellipsis to the end. The
full content will then be displayed in a tooltip by hovering
over the cells.

Signed-off-by: Samir Aguiar <samirjaguiar@gmail.com>