Job: rename args to config
This makes the, previously named args parameter, more accurately a
configuration that can be suplied to a job.  It makes it clearer that
this dictionary doesn't have a dependency on the command line parser
(like it did before becoming a dictionary).

It's expected that the job configuration can be loaded by other from
other sources (JSON, YAML) that can be converted to a dictionary, or
even expressed as a instance in the job definition.

Signed-off-by: Cleber Rosa <crosa@redhat.com>