Job: allow a job to define multiple test suites
It's a common pattern to have to apply different conditions on
different sets of tests.  So far, the only way to do that is to
basically run different jobs.

This patch starts to add the concept of multiple tests suites.  At
this point, it introduces properties to support multiple test suites,
but doesn't change any of the behavior.  This is intentional as we'll
try not to break plugins that depend on the Job behavior.

Signed-off-by: Cleber Rosa <crosa@redhat.com>