Fix filtering and popover in HTML report
After commit 0d8b50e the popover of the info column was broken
due to the HTML title attribute being added. This column doesn't
need that tooltip as it already has a bootstrap popover applied
to it by __init__.py when filling this column.

Also, after commit d6f80ad the ability to filter by status was
lost after wrapping status element in divs, so we need to unwrap
them when filling the dropdown. Finally, bring back a CSS style
that was removed but is needed to make the status dropdown look
the same as the page dropdown added by datatables.

Signed-off-by: Samir Aguiar <samirjaguiar@gmail.com>