nrunner: look for external binaries as runners
The job-like behavior of the nrunner, that is, when you run `avocado
nrun reference1 reference2` will resolve test references into
runnables, and subsequently tasks.

Up to this point, the runner itself, that is the Python module
`avocado.core.nrunner` would be used to run those tasks.  With this
change, Avocado will look for other runners, specific to the kind of
runnable a task has.

This allows, say, an `avocado-runner-glib` to be used to run
tasks/runnables of kind glib.

Likewise, `avocado-runner-avocado-istrumented`, to be introduced
shortly, will handle tasks/runnables of kind `avocado-instrumented`.

Signed-off-by: Cleber Rosa <crosa@redhat.com>