Revert "selftests: Improve avocado_abs command"
The commit this reverts was added to handle the unittest execution
within the context of coverage reporting.  But, it breaks the test
execution when the AVOCADO variable can not be splitted into more than
one component.

Later, a skip for this test was added when running within the context
of coverage reporting, so reverting this would seem logical even if it
was not breaking a test.

This reverts commit f87bab3513459415ba1088ec637c56d952132022.

Signed-off-by: Cleber Rosa <crosa@redhat.com>