nrunner: look for Python modules and other binaries as runners
The job-like behavior of the nrunner, that is, when you run `avocado
nrun reference1 reference2` will resolve test references into
runnables, and subsequently tasks.

Up to this point, the runner itself, that is the Python module
`avocado.core.nrunner` would be used to run those tasks.  With this
change, Avocado will look for runners specific to the kind of runnable
a task has.

This allows a Python module named `runner_foo` under `avocado.core`,
that is, importable as `avocado.core.runner_foo`, to be used for
running runnables of kind `foo`.

Likewise, it allows, an `avocado-runner-glib` executable file anywhere
in the searchable "$PATH" to be used to run tasks/runnables of kind glib.

Signed-off-by: Cleber Rosa <crosa@redhat.com>