Tags: extend the supported character set allowing dashes
Some use cases, for instance on QEMU, will use tag values as a default
value for some key characteristics of the test.  For instance:

   :avocado: tags=arch:x86_64

Is already used for determining the architecture specific aspect of a
test, and will that value set if a parameter was not explicitly set.
But, due to the limited character set supported in tags, some values
can not be set.  For instance:

   :avocado: tags=machine:s390-ccw-virtio

Cannot be properly recognized without dashes being supported.

Signed-off-by: Cleber Rosa <crosa@redhat.com>