Makefile: show help message as the default target
Because of the addition of the Makefile.include file, the first
target was changed to the first on Makefile.include, while it
should have been the "all" target that shows the help message.

Let's change the order of the include to restore the behavior
of showing the help message when running just `make`.

Signed-off-by: Cleber Rosa <crosa@redhat.com>