avocado.utils.process: function for getting matching output
It's definitely a common pattern to run commands, and if the output
matches a given expected pattern, have that output.  On a previous
discussion, I pointed out that:

  "I think this could be introduced as an avocado.utils.process
  utility instead, one that takes a command and a pattern. I don't
  think this justifies a new module (avocado.utils.perf) and the
  hardcoded perf list list command. If, the perf list command output
  were to be parsed into a proper data structure, then I think it'd
  make sense to have a perf module."

What I meant there, it's that a generic utility function could be
added, not one that runs "perf list".  This makes the function
introduced here to be a generic function.  Users for this function
should simply start using:

   get_command_output_matching("perf list", "pattern")

Reference: https://github.com/avocado-framework/avocado/pull/3108
Signed-off-by: Cleber Rosa <crosa@redhat.com>