avocado.utils.ssh: unwind parameters
The current class init signature have two parameters which are tuples.
While there is a valid motivation for this style (such as the first
being similar to what's given to sockets), it was clear that they are
flexible.

Let's unwind the parameters from the tuples, creating properly named
parameters, and while at it, do some basic sanity checking for the
options and the generated ssh command line.

Signed-off-by: Cleber Rosa <crosa@redhat.com>