avocado.utils.ssh: support password based authentication
This adds password based authentication, and bases itself on the
SSH_ASKPASS feature of openssh clients.  Because password is usually
read from the tty, a new session is started and environment variables
are set to a script that will return the password.

This is only used during the establishment of the master session, and
all other ssh executions (used to run the actual commands) will just
rely on the already connection session.

Signed-off-by: Cleber Rosa <crosa@redhat.com>