nrunner: move command line composition into the module
This introduces `runnable_to_command_line_args()` which takes care of
serializing the attributes of a runnable into command line args.
The same should be possible for serializing a runnable into a "recipe".

The idea is to expand the logic and coverage, so that it's fully
tested and reusable, instead of only available in the plugin itself.

Signed-off-by: Cleber Rosa <crosa@redhat.com>