nrunner: support complex runnable arguments via base64 encoded args
Sometimes the arguments to be passed to runnables can not be given on
the command line as simple strings.  Even simple strings can't always
be used because they'll be parsed by the runner command line parser
itself.

To allow for more flexible argument passing, this allows a escaping
prefix "base64:" which will be decoded before being passed into the
Runnable constructor.

One of the functional tests checks that behavior, while the other
serves to make the point that no encoding is necessary in recipe files
as they are JSON and can hold natively the content we need so far.

Signed-off-by: Cleber Rosa <crosa@redhat.com>