nrunner: make use of keyword arguments from the command line
So far, the command line parser has not made proper use of of the
kwargs attribute in the Runnable class.  This adds proper command line
parsing, and makes sure that the key=val arguments given on the
command line are properly reflected on the Runnable instances.

It's up to a runner to map the kwargs in the runnable to their own
need.

Signed-off-by: Cleber Rosa <crosa@redhat.com>