nrunner: avoid confusion with positional arguments at parsing time
From argparse's perspective, positional arguments are the ones that
are not preceeded by a dash or double dash.

The runnable-run "-a|--args" argument is not a positional argument per
se, but is usually passed along to the runner as positional arguments.
Let's change its help messages to "simple arguments" given the fact
that "key word based arguments" will receive an overhaul next, and are
in fact (from argparse's perspective) positional arguments to
runnable-run.

Signed-off-by: Cleber Rosa <crosa@redhat.com>