nrunner: use kwargs in ExecRunner as the process environment
The two primary ways to influence a process execution are command line
arguments and environment variables.  The runnable's args attribute
(and the matching -a|--arg command line options) already allows one
to add to the arguments of the process to be executed.

The keyword arguments were not being used in the ExecRunner (and
derived).  Given that key=val map nicely to environment variables,
let's forward those, if set, to the environment on executed processes.

Signed-off-by: Cleber Rosa <crosa@redhat.com>