nrunner: support JSON encoded keyword arguments
Some input data to runnables is more complex than can be expressed in
in the command line as strings.  Let's support JSON encoded keyword
arguments for those use cases.

Currently this is limited to values of key word arguments.  If the
need arises, we can always expand the support.

Because up to this point it was not possible to serialize complex data
types in the keyword arguments as command line arguments, there was no
support for that.  Now, there's basic serialization support.

Signed-off-by: Cleber Rosa <crosa@redhat.com>