nrunner: add tags as first class attributes for runnables
While tags are used by the job, to determine if a test should be
eligible to be executed based on what the user has requested, tags are
also reused, that is, made available to, Avocado instrumented tests.

Let's make the tags available in the runnables, so that runners can
use them appropriately, that is, make them to the tests (to "executing
runnables", to be more precise).

Signed-off-by: Cleber Rosa <crosa@redhat.com>