sysinfo: don't use a shell to mask out errors
When collecting command output, sysinfo will currently use a shell
which is unnecessary, will mask out errors and will put the wrong
content on files on error conditions.

Example is running `avocado sysinfo` on a system without brctl.

   $ cat sysinfo-2019-10-01-14.03.39/pre/brctl\ show
   /bin/sh: brctl: command not found

With this change, this a sample output of the `avocado sysinfo`
execution:

   $ avocado sysinfo
   avocado.sysinfo: Commands configured by file: /home/cleber/src/avocado/avocado/etc/avocado/sysinfo/commands
   avocado.sysinfo: Files configured by file: /home/cleber/src/avocado/avocado/etc/avocado/sysinfo/files
   avocado.sysinfo: Profilers configured by file: /home/cleber/src/avocado/avocado/etc/avocado/sysinfo/profilers
   avocado.sysinfo: Profiler disabled
   avocado.sysinfo: Could not execute "brctl show": [Errno 2] No such file or directory: 'brctl' (brctl show): 'brctl'
   avocado.sysinfo: Logged system information to /home/cleber/sysinfo-2019-10-01-14.20.16

And the 'brctl show' file is not generated, instead of being generated
without the wrong content.

Signed-off-by: Cleber Rosa <crosa@redhat.com>