Packit Initial Support
This is an initial attempt to implement Packit as a service to
build packages on pull requests.

The goal is to not have changes that will cause breakages to
packaging, and thus, releases should be smoother.  Also, if users
choose to, they should be able to have RPMs for all development
versions at any time too.

Signed-off-by: Cleber Rosa <crosa@redhat.com>