avocado.utils.asset: generously increase the lock acquisition timeout
Under systems with heavy load, and multiple tests contending for
access to an asset file, the test can error because of the short
timeout.

This is a simple workaround that favors less false negatives and
trades it off with a possibly longer test execution time.  The
complementary and definitive solution, though, is to avoid
recalculating the hash if a "verified" flag is in place and it's newer
than the asset file itself.

Reference: https://trello.com/c/WdnjY4n7
Signed-off-by: Cleber Rosa <crosa@redhat.com>