Runner: reduce coupling on run_suite()
run_suite() is really *the* interface used by the job, but so
far it has not been defined as what its responsibilities are.
It ends up doing way too much, and gets involved in the job's
own matters.

The end goal is to have a better interface, but for now, let's at
least make what we have more clear.

Signed-off-by: Cleber Rosa <crosa@redhat.com>