nrunner: give time_start and time_end information
The avocado.core.test.Test class provides a large number of
information on each state message that gets sent to the runner.  Among
those are some related to the execution time of each test.  That
information is consumed by the job and makes its way into the test
results.

Let's start providing information on the various nrunner based
runners, taking the time (pun not intended) to think if those are
really necessary or the responsibility of the runner.  For instance,
the "time_elapsed" information is probably best calculated by the job
and not by each individual runner.

Signed-off-by: Cleber Rosa <crosa@redhat.com>