Fix and standardize HTML report tooltips
The HTML plugin currently uses three kinds of tooltips: a popover
created in Python code for the Info column, a popover created in
Javascript code for elements with the title attribute and a custom
tooltip added in d6f80ad08.

This commit unifies them using Bootstrap's built-in popovers,
which are created in Javascript code. Those popovers are only added
for cells that are visible (i.e. after a datatables redrawn) and
for those with truncated text, which should be slighly faster.
This should avoid the need to maintain a custom tooltip and keep
the code more consistent.

Signed-off-by: Samir Aguiar <samirjaguiar@gmail.com>