Docs: improve the docstring about the parser argument
The command line parser is really the entire "avocado" command line
parser, and not limited to the "test running" aspects.  The original
text is not wrong, but a bit misleading.

While at it, let's reduce the possible confusion that may be caused by
the `avocado.core.parser` module having two similar classes, with very
different origins, and give the right type of the parser argument to
the configure method.

Signed-off-by: Cleber Rosa <crosa@redhat.com>