Test: use temporary log location if one is not given explicitly
An Avocado test can be executed not only by an Avocado runner, but by
other runners such as a plain unittest runner.  In situations such as
the unittest runner, a "base_logdir" argument is not given, and a
directory for the the test results is created in a permanent location.

Let's allow the a test to be instantiated without an explicit
base_logdir and without creating a results directory in a permanent
location (such as the default ~/avocado/job-results).

This is a fix for `make check-full` executions, in which the Avocado
tests in selftests/functional/test_nrunner_interface.py would have
their directories created in ~/avocado/job-results) when run by the
unittest runner).

Signed-off-by: Cleber Rosa <crosa@redhat.com>