selftests/run: make check of temp dirs optional
The same check already runs in all check related make targets.  Let's
make this in-between-test execution optional, as a tool to debug which
test is leaving the directory behind.

Because stopTest() gets run before a test tearDown(), we attempt to
run the test tearDown() so that test's clean ups gets reflected in
the temporary dir check.

Signed-off-by: Cleber Rosa <crosa@redhat.com>