resolver: create standard utility function to split references
Given that most of the behavior of loader and resolver module
should be similar, let's revisit some of the loader features,
refactor and document them, and use them on the resolver code
too.

The initial step is to make the reference split common across both
loader and resolver.  A new module is introduced here to host these
reference handling utilities.

Signed-off-by: Cleber Rosa <crosa@redhat.com>