nrunner: support for serialization of recipes for runnables
The nlist command is a simple wrapper aroung the resolver.  And the
resolver basically returns a result condition for each reference
given, and, if successful, a collection of Runnable instances.

Runnables are intended to be easily serialized as "recipes" files (for
lack of a better name), which are currently JSON based.

With this option, which is also experimental just like the rest of
nrunner and may go away or change name, one can see and preserve the
result of a resolution.  The resulting recipes should be executable by
`avocado-runner-$(kind) runnable-run-recipe $(path_to_recipe)`.

Signed-off-by: Cleber Rosa <crosa@redhat.com>