Implements the assets fetch functionality during the JobPreTests
This implementation adds the functionality of assets fetch command
JobPreTests. This way, all assets are fetched prior to the tests starts.
If the asset fails to be downloaded, it will try again during the test
execution. This is a note for the future implementation of a decorator
that skips the test if an asset is not available.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>