Implements FetchAssetJob plugin to fetch assets during JobPreTests
This implementation adds the functionality of assets fetch command
during JobPreTests. This way, all assets are fetched prior to the
tests starts. If the asset fails to be downloaded, it will try again
during the test execution. This is a note for the future
implementation of a decorator that skips the test if an asset is
not available.

Also, it fixes test_default_enabled_plugins to look at correct output
line.  With the inclusion of assets fetch feature on JobPreTests, log
file line moved from second to third.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>