docs: new improved documentation
The current Avocado documentation is outdated and a bit confusing for
new users. Documentation is one of the starting points for understanding
how Avocado works and deciding whether to use it or not. This change
splits the current documentation into guides and chapters to improve the
readability and reduce the learning curve. During this process I tried
to change as minimum as possible the current content and added only few
sections mainly on Contributor's Guide.

Signed-off-by: Beraldo Leal <bleal@redhat.com>