Yaml to Mux: move tests to the plugin
Some tests seem use the varianter features to make sure there's cross
coverage with other features.  Because we don't have a non-optional
varianter implementation, those tests are better living with the
plugin powering the feature itself.

In some cases, because the variants are kind of an extension of the
basic test, we can optionally run it the plugin is available (see
the "GetData" tests).

Signed-off-by: Cleber Rosa <crosa@redhat.com>