Travis CI: add an s390x job with Python 3.7
This adds an s390x job, which limits the plugins that will be enabled,
excluding the ones that need Fabric, which is very hard to install
from pip and unavailable on many distros from native packages.

It doesn't seem to be necessary to run the same jobs on different
architectures for all Python versions, so let's stick with 3.7.

Signed-off-by: Cleber Rosa <crosa@redhat.com>