vmimage: allow the qemu-img binary to be overriden
Avocado currently searches for the qemu-img binary on the system,
picking the first one available (respecting the PATH environment
variable and looking at common places).

In some specific circurnstances, such as when testing QEMU itself,
it's desirable to use a matching qemu-img binary.  Let's allow this
configuration to be done at the module level, to be less intrusive on
the API, so that the vast majority of vmimage users don't have to
care about it.

Signed-off-by: Cleber Rosa <crosa@redhat.com>