SPEC: add support for EL 8 packages
Due to the lack of some packages on EL and EPEL 8, a few plugins are
not enabled (when compared to the Fedora builds).

Because only EL 8 is supported (due to Python 3 only requirements),
we're not enforcing any version check.  This may be needed in the
future, but it's probably a problem for the not so immediate future.

Signed-off-by: Cleber Rosa <crosa@redhat.com>