Travis CI: disable parallel lint on non amd64
All failures I've seen so far on the non amd64 jobs are due to failed
lint checks running in parallel.  Some sort of race condition between
lint processes seems to be happening. Let's disable parallelization on
those environments until a proper solution is found and applied.

BTW, inspekt will, by default, run its checks in parallel, so the
value given actually controls the number of processes used.  In this
case, it will be limited to one, "disabling" the parallelization
feature.

The name of the variable was chosen to be different than the
previously used one, because this one makes it clear that it contains
the number of parallel jobs, and not a enabled/disabled flag as it
seems to have been intended on the previous variable.

Signed-off-by: Cleber Rosa <crosa@redhat.com>