safeloader: do not require to find module at parent path
The safeloader attempts to find modules at a parent location
in an attempt to find classes that may inherit from avocado.Test.
But, it should not depend on being able to find those modules,
and simply bail out if they're are not a parent location.

Test code and bug found by Nicolas Brignone.

CC: Nicolas Brignone <nbrignon@redhat.com>
Signed-off-by: Cleber Rosa <crosa@redhat.com>