Add an initial pylintrc file to be used with IDEs
While developing, some IDEs will report lint problems that are not
covered by selftests/inspekt-lint.sh. The developer may disable the
message in the source code without knowing it is not covered by the
lint check. The pylintrc can be used to avoid this kind of mismatch.
The messages enabled/disabled in the initial file are those listed
on selftests/inspekt-lint.sh.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>