vmimage: avoid taking snapshots when only the base image is needed
The avocado.utils.vmimage library currently takes non-lazy approach,
calling `get()` when it's instantiated, even though it has some
provision for acting lazily with its `path` and `base_image`
properties.

This change keeps the behavior of favoring a snapshot image when
`get()` is called, because most of the times we expect that users want
a image they can interact with and throw away.  But, when doing other
operations, such as downloading the images, it's totally unnecessary
to create snapshots.

Let's avoid duplicity on the action at instantiation time, allowing
users to either call `get()` or `download()`, or have access to the
snapshot by accessing the `path` property or the base image by
accessing the `base_image` property.

Signed-off-by: Cleber Rosa <crosa@redhat.com>