Refactor Asset class to prepare for new features.
Some new requirements on assets need a more flexible Asset class.
The intent of this patch is to refactor some parts of the Asset class,
making it simple and improving the flexibility for future extensions.

The option to save an asset on `by_name` directory was removed. All
assets are now saved on `by_location` directory. Directory hash is
based on asset location. This makes it easy to implement new features,
like the `find` and `list` command line options.

Added some checks on class constructor to avoid breaking the class API.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>