selftests: replacing inspekt with pylint
Even with the debug, log and verbose options, inspekt output is not so
verbose. Besides that, inspekt was missing some errors because was not
looking recursively into modules and also having problems to find some
modules when running in our CI. This is an attempt to fix thoses issues
replacing it with pure pylint.

Signed-off-by: Beraldo Leal <bleal@redhat.com>