selftests/unit/test_utils_cloudinit.py: use automatically aloccated port
Instead of trying to find a free port on our code, let the socket system
code allocate a port.  Because the server will be running, and the socket
bound, it's possible to get the port number with getsockname().

This fixes an `OSError: [Errno 98] Address already in use` that used to
appear on Travis-CI from time to time.

Signed-off-by: Cleber Rosa <crosa@redhat.com>