YAML Loader: behave when empty references are given
A loader's discover() method can be passed a None reference, which is
a valid scenario, and the given loader may for instance return all
builtin/default tests.

In the case of the YAML Loader, this references is expected to be a
path to YAML file, so it should bail out early when None is given.

This fixes the error that can be seen on the command line similar to:

  Test discovery plugin <avocado_loader_yaml.YamlTestsuiteLoader
  object at 0x7f544a6c62e8> failed: expected string or bytes-like object

Signed-off-by: Cleber Rosa <crosa@redhat.com>