avocado.core: data_dir: Introduce the get_job_results_dir()
Currently data_dir.get_resultsdir() returns the path to the job
results directory although the name does not clearly state
that. So the first change here is renaming it to get_job_results_dir().

Its implementation also changed so that:
- base logs directory parameter is optional: uses the default value
  from settings otherwise
- uses the idea of "job reference" rather can "job id" since in some
  cases it is not being passed an id properly saying
- on passing a valid path to the job logs directory it checks if
  the 'id' file exists
- on passing a valid file path it checks it is an 'id' file
- removed the get_id() from data_dir. Besides the name saying not
  much, it was only used in get_resultsdir(). So its logic was
  added in get_job_results_dir()

The replay and diff plugins - the only clients of get_resultsdir() -
were changed to use get_job_results_dir(). This refactor reduced
the code in those plugins.

Finally it was added test cases to get_job_results_dir().

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>