nrunner: job compliant test runner based on the nrunner
This is still an experimental implementation, that can not be accessed
by users of the "avocado" command line application.  A sample job
script is provided.

The goal of including this plugin at this stage is similar to the
overall goal of including the other nrunner pieces: to allow for
incremental feature implementation in smaller batches, until it's
considered stable and no longer experimental.

Signed-off-by: Cleber Rosa <crosa@redhat.com>