[RFC] nrunner requirements: work directory
The avocado.Test class provides tests with a "workdir", a temporary
directory that will be created by the test runner.  Tests are then
free to create any content that they seem fit there, and they will,
by default, be removed once the test is over.

The concept of "requirements" has been maturing in many discussions,
including around the Asset fetcher.  This is a very simplistic idea
that a temporary work directory may just as well be a test requirement
too.

It's anticipated that different test types will have different requirements,
and the very next one that is similar to this is probably a "tests'
result directory", one that will have its contents persisted, instead
of removed.

Internally, at this point, it's not properly defined as a "test
requirement" because the concept has not been introduced.  Other point
not covered here is that idea that different requirements may have
different "agents" to fulfill them.

Still, I hope that this helps the discussion to move forward.

Signed-off-by: Cleber Rosa <crosa@redhat.com>