avocado.utils.network.is_port_free: default to checking all interfaces
One of the reasons being a simplified API, which should allow users to
simply do:

   >>> network.is_port_free(8080)

The second reason is that lgtm.com complains that binding to all
interfaces is a problem, and it probably does that by looking at the
empty string being used.  Because we know what we're doing, that we
may really want to check port availability on all interfaces, this use
case is justified.  Respecting the address parameter given, though,
seems more correct.

Signed-off-by: Cleber Rosa <crosa@redhat.com>