selftests: acknowledge the use of stderr for its natural purpose
There are a number of selftests that assume that the stderr should
never contain anything.  In truthfulness, that is more appropriate
to the stdout, once the stderr can produce content even when a
silent operation is requested.

On some other cases, the intention of stderr is mixed with that of the
stdout, like attempting to parse JSON from stdout and from stderr.
The content that goes to stderr can include any type of error
messages, and will not conform or be part of the intended JSON
produced by the JSON result plugin.

This fixes the following errors such as:

   ERROR: test_output_compatible_setup (selftests.functional.test_output.OutputPluginTest)
   ----------------------------------------------------------------------
   Traceback (most recent call last):
     File "/builddir/build/BUILD/avocado-a257e52c24b41976ce5ca8f0159d8148f273f0e9/selftests/functional/test_output.py", line 332, in test_output_compatible_setup
       json.loads(output)
     File "/usr/lib64/python3.6/json/__init__.py", line 354, in loads
       return _default_decoder.decode(s)
     File "/usr/lib64/python3.6/json/decoder.py", line 342, in decode
       raise JSONDecodeError("Extra data", s, end)
   json.decoder.JSONDecodeError: Extra data: line 27 column 1 (char 1038)

Signed-off-by: Cleber Rosa <crosa@redhat.com>