config: adding defaults file
This file is experimental. The idea is to have something that is not
changed by the user, only by the developers. For now, this is a conf/ini
file, but we will migrate this to a python file soon. This will not
affect the current behavior, since that is only being parsed by
future/settings.py

Signed-off-by: Beraldo Leal <bleal@redhat.com>