future/settings: Migrate sysinfo to use the new future/settings API
As part of the effort to remove default values scattered around the
code, this change only migrates the sysinfo option to use the new
module. This will associate the --sysinfo command line option with the
enabled configuration of the sysinfo.collect section in the
configuration file. If this gets accepted, I will start migrating all
the options for this new model.

Signed-off-by: Beraldo Leal <bleal@redhat.com>