Don't trust HTML in jinja template
The info column in the HTML report was marked as "safe" in the
jinja template, meaning that any HTML within it would be rendered
and not escaped. However, since the contents of this column can be
outside the control of the tests (e.g. an error message returned
by a web service), it's best not to trust it.

The same happens with the sysinfo panel where we used HTML to
display links to files that could not be read when generating the
report. In this case, however, we can keep the same behavior while
still avoiding unescaped HTML content.

Signed-off-by: Samir Aguiar <samirjaguiar@gmail.com>